stasher: GET request does not acquire a lock when not in quorum.
For local client connections, GET requests should not acquire
object locks when the repository is not in quorum. This is because
this can happen because the slave is syncing itself from the
master, and the repo sync needs to acquire object locks in order
to resynchronize the local repository copy. The lock object is
already a nullable ptr, so what we need to do here is to factor
out lock acquiisition into a more general method that either
acquires a lock for a GET request, or drops one, depending on
the current state, and call it for all the objects in the
get queue when the quorum state changes.