stasher: fix race condition with connection establishment.
When repopeerconnectionObj goes out of scope and gets destroyed, updateThread
gets started for the purpose of invoking recalculate().

When recalculate() changes this node's master, startmaster() gets invoked
to start the new master controller.

When repopeerconnectionObj to the master goes away, if a new connection
gets established to the same node, the new repopeerconnectionObj invokes
installpeer(), and if installpeer's call to recalculate() selects the same
master, since the master never actually gets changed, startmaster() does not
end up getting called.

We'll fix this in installpeer(). If it sees that this node's current master
is supposedly the newly-connected peer, something is wrong. Rejecting the
peer connection isn't ideal, but it allows the cluster to eventually recover
by itself.

Also: clean up formatting. Make an internal data type private.