stasher: suppress errant quorum announcement from master controller.
Master controller starts with an empty list of slaves, and it errantly
announces a quorum.

Postpone the quorum announcement until the master controller thread gets the
initial cluster state, from the cluster notification callback.

Replace the initial announcement with a fake not-in-quorum announcement, and
when the initial quorum state gets received, announce it.