Extend subscriptions to sub-hierarchies.
Subscribing to a hierarchy results in notifications of objects that are
changed in the hierarchy and all sub-hierarchies.

Use x::hier to register all subscribed hierarchies. When the subscription
request for a hierarchy comes in, in addition to doing what's happening now,
also register the hierarchy in the x::hier container.

Remove the code that notifies the immediate hierarchy of a changed object
to searching the x::hier container, then walking up the hierarchy, and
notifying everyone.