Updated managed subscriptions to sub-hierarchy scope.
Now that the underlying client subscriptions cover the entire hierarchy and
all the sub-hierarchies, the managed subscription code needs to be updated
only to scan he entire hierarchy and all subhierarchies initially, and the
subscription update is going to take care of the ongoing updates.

Add a list of hierarchies to the implementation object, and when the connection
is made, prime the list with the top-level hierarchy.

The code that requests the directory listing of the hierarchy is factored out
into req_next_hierarchy() that takes the first hierarchy name off the list,
and calls enumerated() if the list is empty.

The directory results callback removes the names of all the subhierarchies
from the directory contents results, and adds them to the hierarchy list,
before calling update().

Remove the code that skips the names of all the subhierarchies from update(),
this is done before it gets called. If there are no more directory entries
to process, do not call enumerated(), call req_next_hierarchy(), instead.